pyproject.toml
[project]
name = "vybe-virtual-tryon"
version = "0.1.0"
description = "FastMCP server for Vybe Virtual Try-On using Replicate API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "replicate",
    "python-dotenv",
    "uvicorn",
    "httpx",
    "authlib>=1.3.0",
    "python-jose[cryptography]>=3.3.0",
    "itsdangerous>=2.1.2",
    "python-multipart>=0.0.6",
    "starlette>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

# Remove build-system since this is a standalone script, not a package
# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]